Implement -v.

+++ compress.h
@@ -1,7 +1,7 @@
 typedef struct
 {
     const char *path, *name_in, *name_out;
-    size_t sz, sd;
+    unsigned long long sz, sd;
 } file_info;
 
 typedef int(compress_func)(int,int,file_info*restrict);
